[bugfix] 1.番剧下载丢文件 2.后台恢复卡数据更新 3.优化番剧下载逻辑，解决一部分 parsing timeout 问题 4.优化番剧下载弹窗（新增剧集会默认位于第一个）5.下载失败支持点击重试 [feature] 1.新增可选 nomedia 模式 2.新增最大下载任务数配置 3.新增h265下载编码与配置

+++ app/src/main/java/com/heyanle/easybangumi4/cartoon/story/download/runtime/CartoonDownloadDispatcher.kt
@@ -30,8 +30,8 @@ class CartoonDownloadDispatcher(
     // 调度统一给主线程调度
     private val scope = MainScope()
     private val executor = ThreadPoolExecutor(
-        cartoonDownloadPreference.downloadMaxCountPref.get(),
-        cartoonDownloadPreference.downloadMaxCountPref.get(),
+        cartoonDownloadPreference.downloadMaxCountPref.get().toInt(),
+        cartoonDownloadPreference.downloadMaxCountPref.get().toInt(),
         50,
         TimeUnit.SECONDS,
         LinkedBlockingQueue<Runnable>(),
